Add resizable traits
