+++ src/Models/Post.php
@@ -6,11 +6,13 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Auth;
 use TCG\Voyager\Facades\Voyager;
+use TCG\Voyager\Traits\Resizable;
 use TCG\Voyager\Traits\Translatable;
 
 class Post extends Model
 {
     use Translatable;
+    use Resizable;
 
     protected $translatable = ['title', 'seo_title', 'excerpt', 'body', 'slug', 'meta_description', 'meta_keywords'];
 
@@ -51,20 +53,5 @@ class Post extends Model
     public function category()
     {
         return $this->hasOne(Voyager::modelClass('Category'), 'id', 'category_id');
-    }
-
-    /**
-     *   Method for returning specific thumbnail for post.
-     */
-    public function thumbnail($type)
-    {
-        // We take image from posts field
-        $image = $this->attributes['image'];
-        // We need to get extension type ( .jpeg , .png ...)
-        $ext = pathinfo($image, PATHINFO_EXTENSION);
-        // We remove extension from file name so we can append thumbnail type
-        $name = rtrim($image, '.'.$ext);
-        // We merge original name + type + extension
-        return $name.'-'.$type.'.'.$ext;
     }
 }
