+++ src/Traits/Resizable.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace TCG\Voyager\Traits;
+
+trait Resizable
+{
+    /**
+     *   Method for returning specific thumbnail for model.
+     */
+    public function thumbnail($type)
+    {
+        // We take image from posts field
+        $image = $this->attributes['image'];
+        // We need to get extension type ( .jpeg , .png ...)
+        $ext = pathinfo($image, PATHINFO_EXTENSION);
+        // We remove extension from file name so we can append thumbnail type
+        $name = rtrim($image, '.'.$ext);
+        // We merge original name + type + extension
+        return $name.'-'.$type.'.'.$ext;
+    }
+}
